test(azdo): split work item tag tests
